Code cleanup
- In previous PRs I have noticed that some overrides are bit surprising which
  sometimes lead to mistakes. Also refactoring is safer when compiler screams
  when I remove/add something with unexpected consequences.
- Added @Override annotations
- Added missing serialVersionUID numbers
- Added missing @Deprecated annotations
- Removed redundant modifiers
- Removed redundant classcasts
- Variable declarations of local variables moved to their first usage.

+++ idlj/src/main/java/com/sun/tools/corba/ee/idl/constExpr/Equal.java
@@ -52,7 +52,7 @@ public class Equal extends BinaryExpr
           value (new Boolean (l.doubleValue () == r.doubleValue ()));
         else
           //daz          value (Boolean.valueOf (l.longValue () == r.longValue ()));
-          value (Boolean.valueOf (((BigInteger)l).equals ((BigInteger)r)));
+          value (Boolean.valueOf (((BigInteger)l).equals (r)));
       }
     }
     catch (ClassCastException e)
